Change Note: + Features Class Introduced + Comment +- Bugs Fixes
